While investigating an issue with queries that use the `timestamp()` function with Mimir's chunks streaming (https://github.com/grafana/mimir/pull/5370), we discovered that the current special case handling for this function creates a new series iterator for each series for each time step in the query, as well as a new `MemoizedSeriesIterator` for each time step.

Not only does this break Mimir's chunks streaming (because creating a second iterator for a series requires reading chunks that we've already read past), it also turns out to be quite inefficient. Changing the special case for `timestamp()` to create and retain one iterator per series for the whole query has the following impact on the benchmark added in this PR:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql
                                                   │  before.txt  │              after.txt              │
                                                   │    sec/op    │    sec/op     vs base               │
RangeQuery/expr=timestamp(a_one),steps=1-10           10.74µ ± 1%   11.84µ ±  7%  +10.19% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=10-10          25.56µ ± 1%   12.54µ ±  7%  -50.95% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=100-10        271.63µ ± 1%   36.07µ ±  6%  -86.72% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=1000-10       8798.2µ ± 1%   221.0µ ± 11%  -97.49% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=1-10           48.00µ ± 1%   36.20µ ±  9%  -24.58% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=10-10         187.21µ ± 1%   56.89µ ± 10%  -69.61% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=100-10        2600.1µ ± 1%   290.3µ ±  8%  -88.84% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=1000-10       89.292m ± 2%   2.562m ±  7%  -97.13% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=1-10       416.8µ ± 1%   322.0µ ± 20%  -22.74% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=10-10     1816.6µ ± 1%   562.5µ ±  7%  -69.04% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=100-10    26.112m ± 1%   2.542m ± 20%  -90.26% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=1000-10   896.10m ± 1%   22.67m ±  1%  -97.47% (p=0.002 n=6)
geomean                                               1.304m        227.2µ        -82.58%

                                                   │  before.txt   │              after.txt              │
                                                   │     B/op      │     B/op      vs base               │
RangeQuery/expr=timestamp(a_one),steps=1-10           7.506Ki ± 0%   7.013Ki ± 0%   -6.57% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=10-10         12.782Ki ± 0%   7.787Ki ± 0%  -39.08% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=100-10         68.79Ki ± 0%   15.76Ki ± 0%  -77.09% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=1000-10       994.03Ki ± 0%   96.04Ki ± 0%  -90.34% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=1-10           21.54Ki ± 0%   22.25Ki ± 0%   +3.26% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=10-10          47.25Ki ± 0%   28.25Ki ± 0%  -40.21% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=100-10        339.33Ki ± 0%   89.54Ki ± 0%  -73.61% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=1000-10       6996.4Ki ± 0%   703.0Ki ± 0%  -89.95% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=1-10       169.1Ki ± 0%   181.8Ki ± 0%   +7.57% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=10-10      398.3Ki ± 0%   239.4Ki ± 0%  -39.90% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=100-10    3044.3Ki ± 0%   826.3Ki ± 0%  -72.86% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=1000-10   65.363Mi ± 0%   6.540Mi ± 0%  -89.99% (p=0.002 n=6)
geomean                                               278.5Ki        98.03Ki       -64.80%

                                                   │  before.txt   │             after.txt              │
                                                   │   allocs/op   │  allocs/op   vs base               │
RangeQuery/expr=timestamp(a_one),steps=1-10             152.0 ± 0%    147.0 ± 0%   -3.29% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=10-10            233.0 ± 0%    174.0 ± 0%  -25.32% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=100-10          1192.0 ± 0%    447.0 ± 0%  -62.50% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_one),steps=1000-10        19.886k ± 0%   3.167k ± 0%  -84.07% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=1-10             290.0 ± 0%    285.0 ± 0%   -1.72% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=10-10            624.0 ± 0%    322.0 ± 0%  -48.40% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=100-10          5686.0 ± 0%    710.0 ± 0%  -87.51% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_ten),steps=1000-10       153.082k ± 0%   4.561k ± 0%  -97.02% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=1-10        1.584k ± 0%   1.579k ± 0%   -0.32% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=10-10       4.404k ± 0%   1.672k ± 0%  -62.03% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=100-10     50.081k ± 0%   2.801k ± 0%  -94.41% (p=0.002 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=1000-10   1480.41k ± 0%   13.78k ± 0%  -99.07% (p=0.002 n=6)
geomean                                                4.497k        1.008k       -77.59%

```

One drawback of this approach is that all chunks for all series will be held in memory while evaluating the query, including when using Mimir's chunks streaming. However, this is no worse than the current non-streaming situation. Modifying the PromQL engine to operate per-series in this case would require more complex changes to the PromQL engine code, so I'd like to test how things fare with this change in place, and then come back and revisit this if it proves to be problematic.

Note to reviewers: I'd suggest reviewing each commit separately, as they incrementally build up the changes required.

I'll submit this to `prometheus/prometheus` once I've tested this in a Mimir test environment.