@charleskorn 2023-06-29
> I was wondering if there could be an issue if you have a sample every 1m but a query step = 1s, so for 60 steps we have to lookup the same sample. I don't think it's a problem because `vectorSelectorSingle()` will call `iterator.Seek()` (which will not advance the iterator given the current sample timestamp < step) and then `vectorSelectorSingle()` will look up at the previous sample calling `PeekPrev()`.

Yep, that's my expectation as well. I'll add a test to confirm this though.