@charleskorn 2023-07-20
I've created https://github.com/prometheus/prometheus/pull/12579 to upstream these changes.